fix: remove circular dependencies from transactions
