+++ packages/transactions/src/authorization.ts
@@ -10,7 +10,11 @@ import {
 } from './constants';
 
 import { BufferArray, cloneDeep, leftPadHex, txidFromData } from './utils';
-
+import {
+  TransactionAuthField,
+  serializeMessageSignature,
+  deserializeMessageSignature,
+} from './signature';
 import {
   addressFromPublicKeys,
   createEmptyAddress,
@@ -20,138 +24,24 @@ import {
 } from './types';
 
 import {
-  compressPublicKey,
   createStacksPublicKey,
-  deserializePublicKey,
   getPublicKey,
   isCompressed,
   publicKeyFromSignature,
-  serializePublicKey,
   signWithKey,
   StacksPrivateKey,
   StacksPublicKey,
 } from './keys';
 
 import { BufferReader } from './bufferReader';
+import { MessageSignature } from './common';
 import { DeserializationError, SigningError, VerificationError } from './errors';
-
-export interface MessageSignature {
-  readonly type: StacksMessageType.MessageSignature;
-  data: string;
-}
-
-export function createMessageSignature(signature: string): MessageSignature {
-  const length = Buffer.from(signature, 'hex').byteLength;
-  if (length != RECOVERABLE_ECDSA_SIG_LENGTH_BYTES) {
-    throw Error('Invalid signature');
-  }
-
-  return {
-    type: StacksMessageType.MessageSignature,
-    data: signature,
-  };
-}
 
 export function emptyMessageSignature(): MessageSignature {
   return {
     type: StacksMessageType.MessageSignature,
     data: Buffer.alloc(RECOVERABLE_ECDSA_SIG_LENGTH_BYTES, 0x00).toString('hex'),
   };
-}
-
-export function serializeMessageSignature(messageSignature: MessageSignature): Buffer {
-  const bufferArray: BufferArray = new BufferArray();
-  bufferArray.appendHexString(messageSignature.data);
-  return bufferArray.concatBuffer();
-}
-
-export function deserializeMessageSignature(bufferReader: BufferReader): MessageSignature {
-  return createMessageSignature(
-    bufferReader.readBuffer(RECOVERABLE_ECDSA_SIG_LENGTH_BYTES).toString('hex')
-  );
-}
-
-enum AuthFieldType {
-  PublicKeyCompressed = 0x00,
-  PublicKeyUncompressed = 0x01,
-  SignatureCompressed = 0x02,
-  SignatureUncompressed = 0x03,
-}
-
-export type TransactionAuthFieldContents = StacksPublicKey | MessageSignature;
-
-export interface TransactionAuthField {
-  type: StacksMessageType.TransactionAuthField;
-  pubKeyEncoding: PubKeyEncoding;
-  contents: TransactionAuthFieldContents;
-}
-
-export function createTransactionAuthField(
-  pubKeyEncoding: PubKeyEncoding,
-  contents: TransactionAuthFieldContents
-): TransactionAuthField {
-  return {
-    pubKeyEncoding,
-    type: StacksMessageType.TransactionAuthField,
-    contents,
-  };
-}
-
-export function serializeTransactionAuthField(field: TransactionAuthField): Buffer {
-  const bufferArray: BufferArray = new BufferArray();
-
-  switch (field.contents.type) {
-    case StacksMessageType.PublicKey:
-      if (field.pubKeyEncoding == PubKeyEncoding.Compressed) {
-        bufferArray.appendByte(AuthFieldType.PublicKeyCompressed);
-        bufferArray.push(serializePublicKey(field.contents));
-      } else {
-        bufferArray.appendByte(AuthFieldType.PublicKeyUncompressed);
-        bufferArray.push(serializePublicKey(compressPublicKey(field.contents.data)));
-      }
-      break;
-    case StacksMessageType.MessageSignature:
-      if (field.pubKeyEncoding == PubKeyEncoding.Compressed) {
-        bufferArray.appendByte(AuthFieldType.SignatureCompressed);
-      } else {
-        bufferArray.appendByte(AuthFieldType.SignatureUncompressed);
-      }
-      bufferArray.push(serializeMessageSignature(field.contents));
-      break;
-  }
-
-  return bufferArray.concatBuffer();
-}
-
-export function deserializeTransactionAuthField(bufferReader: BufferReader): TransactionAuthField {
-  const authFieldType = bufferReader.readUInt8Enum(AuthFieldType, n => {
-    throw new DeserializationError(`Could not read ${n} as AuthFieldType`);
-  });
-
-  switch (authFieldType) {
-    case AuthFieldType.PublicKeyCompressed:
-      return createTransactionAuthField(
-        PubKeyEncoding.Compressed,
-        deserializePublicKey(bufferReader)
-      );
-    case AuthFieldType.PublicKeyUncompressed:
-      return createTransactionAuthField(
-        PubKeyEncoding.Uncompressed,
-        deserializePublicKey(bufferReader)
-      );
-    case AuthFieldType.SignatureCompressed:
-      return createTransactionAuthField(
-        PubKeyEncoding.Compressed,
-        deserializeMessageSignature(bufferReader)
-      );
-    case AuthFieldType.SignatureUncompressed:
-      return createTransactionAuthField(
-        PubKeyEncoding.Uncompressed,
-        deserializeMessageSignature(bufferReader)
-      );
-    default:
-      throw new Error(`Unknown auth field type: ${JSON.stringify(authFieldType)}`);
-  }
 }
 
 export interface SingleSigSpendingCondition {
